상점 수정 진행중

+++ src/main/java/kr/fscom/firsens/sys/domain/SYSFileDomain.java
@@ -15,8 +15,9 @@ public class SYSFileDomain {
     @JsonProperty("areaCode")      private String areaCode;                       // 구역코드
     @JsonProperty("imgName")       private String imgName;                        // 이미지 저장명
     @JsonProperty("imgPath")       private String imgPath;                        // 이미지 경로
-    @JsonProperty("imgSize")       private Long imgSize;                          // 이미지 용량
+    @JsonProperty("imgSize")       private int imgSize;                          // 이미지 용량
     @JsonProperty("imgNum")        private int imgNum;                            // 이미지 순번
     @JsonProperty("oriName")       private String oriName;                        // 이미지 원본명
     @JsonProperty("regDate")       private String regDate;                        // 이미지 업로드 일자
+
 }
